Reformat models to a directory
