+++ retail_app/admin.py
@@ -1,17 +1,15 @@
 from django.contrib import admin
 
-from .models import Business, Designer, Product, ProductStock, ProductColor, ProductImage, ProductDescription, ProductDetails, ProductPrice, ProductQuantity
+from .models import (
+    Business,
+    Designer,
+    Product,
+)
 
 admin.site.register(Business)
 admin.site.register(Designer)
 admin.site.register(Product)
-admin.site.register(ProductStock)
-admin.site.register(ProductColor)
-admin.site.register(ProductImage)
-admin.site.register(ProductDescription)
-admin.site.register(ProductDetails)
-admin.site.register(ProductPrice)
-admin.site.register(ProductQuantity)
 
-fields = ['image_tag']
-readonly_fields = ['image_tag']
+
+fields = ["image_tag"]
+readonly_fields = ["image_tag"]
