+++ retail_app/models/BusinessDesigner.py
@@ -1,0 +1,11 @@
+from django.db import models
+
+
+class BusinessDesigner(models.Model):
+    name = models.CharField(max_length=50)
+
+    def __str__(self):
+        return self.name
+
+    class Meta:
+        verbose_name_plural = "Business Designers"
