+++ retail_app/models/Category.py
@@ -1,0 +1,11 @@
+from django.db import models
+
+
+class Category(models.Model):
+    name = models.CharField(max_length=50)
+
+    class Meta:
+        verbose_name_plural = "Categories"
+
+    def __str__(self):
+        return self.name
