+++ retail_app/models/Collection.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+
+class Collection(models.Model):
+    name = models.CharField(max_length=50)
+
+    def __str__(self):
+        return self.name
