+++ retail_app/models/Designer.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+
+class Designer(models.Model):
+    name = models.CharField(max_length=50)
+    site_url = models.URLField()
+
+    def __str__(self):
+        return self.name
