+++ retail_app/models/ProductColor.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+
+class ProductColor(models.Model):
+    color = models.CharField(max_length=50)
+
+    def __str__(self):
+        return self.color
