+++ retail_app/models/Season.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+
+class Season(models.Model):
+    name = models.CharField(max_length=50)
+
+    def __str__(self):
+        return self.name
